Fixed Merging errors. I moved the breadcrumb function to utils I deleted the product-list.js reference being made in product.js I used a main() asynchronous function to allow the call of loadHeaderFooter to be awaited for the searchbar to work in the product.js file.
