+++ src/checkout/index.html
@@ -13,38 +13,87 @@
   </head>
 
   <body>
-    <header class="divider" id="main-header">
-      
-    </header>
+    <header class="divider" id="main-header"></header>
 
     <main class="divider">
       <section class="products">
         <h2>Review & Place your Order</h2>
         <form name="checkout">
           <fieldset>
-              <legend>Shipping</legend>
-              <label class="form-label">First Name<input type="text" name="fname" id="fname" required></label>
-              <label class="form-label">Last Name<input type="text" name="lname" id="lname" required></label>
-              <label class="form-label">Street<input type="text" name="street" id="street" required></label>
-              <label class="form-label">City<input type="text" name="city" id="city" required></label>
-              <label class="form-label">State<input type="text" name="state" id="state" pattern="[A-Z]{2}" required></label>
-              <label class="form-label">Zip<input type="text" name="zip" id="zip" pattern="[0-9]{5}" required></label>
+            <legend>Shipping</legend>
+            <label class="form-label"
+              >First Name<input type="text" name="fname" id="fname" required
+            /></label>
+            <label class="form-label"
+              >Last Name<input type="text" name="lname" id="lname" required
+            /></label>
+            <label class="form-label"
+              >Street<input type="text" name="street" id="street" required
+            /></label>
+            <label class="form-label"
+              >City<input type="text" name="city" id="city" required
+            /></label>
+            <label class="form-label"
+              >State<input
+                type="text"
+                name="state"
+                id="state"
+                pattern="[A-Z]{2}"
+                required
+            /></label>
+            <label class="form-label"
+              >Zip<input
+                type="text"
+                name="zip"
+                id="zip"
+                pattern="[0-9]{5}"
+                required
+            /></label>
           </fieldset>
           <fieldset>
-              <legend>Payment</legend>
-              <label class="form-label">Card Number<input type="text" name="cardNumber" id="cardNumber" minlength="16" pattern="^[0-9]{16}" required></label>
-              <label class="form-label">Expiration<input type="text" name="expiration" id="expiration"></label>
-              <label class="form-label">Code<input type="number" name="code" id="code" pattern="^[0-9]" minlength="3" maxlength="3"></label>
+            <legend>Payment</legend>
+            <label class="form-label"
+              >Card Number<input
+                type="text"
+                name="cardNumber"
+                id="cardNumber"
+                minlength="16"
+                pattern="^[0-9]{16}"
+                required
+            /></label>
+            <label class="form-label"
+              >Expiration<input type="text" name="expiration" id="expiration"
+            /></label>
+            <label class="form-label"
+              >Code<input
+                type="number"
+                name="code"
+                id="code"
+                pattern="^[0-9]"
+                minlength="3"
+                maxlength="3"
+            /></label>
           </fieldset>
           <fieldset class="checkout-summary">
-              <legend>Order Summary</legend>
-              <p class="summary-legend">Item Subtotal(<span id="number-of-items"></span>)</p><span id="subtotal"></span>
-              <p class="summary-legend">Shipping Estimate</p><span id="shipping-estimate"></span>
-              <p class="summary-legend">Tax</p><span id="tax-subtotal"></span>
-              <p class="summary-legend-bold">Order Total</p><span id="order-total"></span>
+            <legend>Order Summary</legend>
+            <p class="summary-legend">
+              Item Subtotal(<span id="number-of-items"></span>)
+            </p>
+            <span id="subtotal"></span>
+            <p class="summary-legend">Shipping Estimate</p>
+            <span id="shipping-estimate"></span>
+            <p class="summary-legend">Tax</p>
+            <span id="tax-subtotal"></span>
+            <p class="summary-legend-bold">Order Total</p>
+            <span id="order-total"></span>
           </fieldset>
-          <input type="submit" name="checkout-btn" id="checkout-btn" value="Checkout">
-      </form>
+          <input
+            type="submit"
+            name="checkout-btn"
+            id="checkout-btn"
+            value="Checkout"
+          />
+        </form>
       </section>
     </main>
