+++ src/css/style.css
@@ -276,7 +276,7 @@ form input {
   font-size: 18px;
   display: flex;
 }
-.breadcrumbs li{
+.breadcrumbs li {
   padding: 5px;
   text-decoration: none;
   list-style-type: none;
@@ -406,7 +406,6 @@ form input {
   display: block;
   margin: 1rem 0px;
   text-align: center;
-  
 }
 
 #product-selected-qty {
