+++ src/product-list/index.html
@@ -16,7 +16,9 @@
       <li class="breadcrumbs_category">
         <a
           href="../product-listing/index.html"
-          class="breadcrumb_link" id="breadcrumb">Product Category</a
+          class="breadcrumb_link"
+          id="breadcrumb"
+          >Product Category</a
         >
       </li>
     </ul>
