+++ src/product_pages/index.html
@@ -11,7 +11,6 @@
     <link rel="stylesheet" href="../css/style.css" />
 
     <script src="../js/product.js" type="module"></script>
-
   </head>
 
   <body>
@@ -23,13 +22,17 @@
       <li class="breadcrumbs_category">
         <a
           href="../product-listing/index.html"
-          class="breadcrumb_link" id="breadcrumb">Product Category</a
+          class="breadcrumb_link"
+          id="breadcrumb"
+          >Product Category</a
         >
       </li>
       <li class="breadcrumbs_detail">
         <a
           href="../product_pages/index.html"
-          class="breadcrumb_link" id="product_page">Product Page</a
+          class="breadcrumb_link"
+          id="product_page"
+          >Product Page</a
         >
       </li>
     </ul>
@@ -44,7 +47,10 @@
           <p class="line-through" id="suggestedRetailPrice"></p>
         </div>
         <p class="product__color" id="productColorName"></p>
-        <label class="product__quantity">Quantity: <select id="product-selected-qty"></select></label>
+        <label class="product__quantity"
+          >Quantity:
+          <select id="product-selected-qty"></select
+        ></label>
         <div class="product-detail__add">
           <button id="addToCart" data-id="">Add to Cart</button>
         </div>
